+++ src/main/java/javaewah/NonEmptyVirtualStorage.java
@@ -1,29 +1,32 @@
 package javaewah;
 
-
 /*
  * Copyright 2009-2012, Daniel Lemire, Cliff Moon, David McIntosh and Robert Becho
  * Licensed under APL 2.0.
  */
 /**
- * This is a BitmapStorage that can be used to determine quickly
- * if the result of an operation is non-trivial... that is, whether
- * there will be at least on set bit.
+ * This is a BitmapStorage that can be used to determine quickly if the result
+ * of an operation is non-trivial... that is, whether there will be at least on
+ * set bit.
  * 
  * @since 0.4.2
  * @author Daniel Lemire
- *
+ * 
  */
 public class NonEmptyVirtualStorage implements BitmapStorage {
   class NonEmptyException extends RuntimeException {
-    private static final long serialVersionUID = 1L;    
+    private static final long serialVersionUID = 1L;
   }
+
   /**
-   * If the word to be added is non-zero, a NonEmptyException exception is thrown.
+   * If the word to be added is non-zero, a NonEmptyException exception is
+   * thrown.
+   * 
    * @see javaewah.BitmapStorage#add(long)
    */
   public int add(long newdata) {
-    if(newdata!=0) throw new NonEmptyException(); 
+    if (newdata != 0)
+      throw new NonEmptyException();
     return 0;
   }
 
@@ -34,7 +37,8 @@ public class NonEmptyVirtualStorage implements BitmapStorage {
    * @see javaewah.BitmapStorage#addStreamOfEmptyWords(boolean, long)
    */
   public int addStreamOfEmptyWords(boolean v, long number) {
-    if(v) throw new NonEmptyException(); 
+    if (v)
+      throw new NonEmptyException();
     return 0;
   }
 
@@ -43,16 +47,17 @@ public class NonEmptyVirtualStorage implements BitmapStorage {
    * 
    * @see javaewah.BitmapStorage#addStreamOfDirtyWords(long[], long, long)
    */
-  public long addStreamOfDirtyWords(long[] data, long start, long number) {
+  public int addStreamOfDirtyWords(long[] data, int start, int number) {
     throw new NonEmptyException();
   }
 
   /**
    * throws a NonEmptyException exception
    * 
-   * @see javaewah.BitmapStorage#addStreamOfNegatedDirtyWords(long[], long, long)
+   * @see javaewah.BitmapStorage#addStreamOfNegatedDirtyWords(long[], long,
+   *      long)
    */
-  public long addStreamOfNegatedDirtyWords(long[] data, long start, long number) {
+  public int addStreamOfNegatedDirtyWords(long[] data, int start, int number) {
     throw new NonEmptyException();
   }
